mapped some OTUs to just genus or family
Open Tree of Life tree entry issues:
The tree I am uploading has a number of species that are apparently not in the Open Tree Taxonomy (OTT). But, they are the right species (from the definitive monograph). There "oughtta be a way" to have these rather than doing what I am currently doing which is just giving them all the genus name (mostly because I don't want to give up since I just put in 15 minutes and apparently if 1 OTU is not found, the whole thing is not useable).

(Update Study #ot_278 via OpenTree API)