Unmap Pseudomonas stutzeri
If we unmap this, we get back the taxonomic group 'Pseudomonas', which has 20,000 or so direct children.  The mapping is probably actually correct, but either the tree is wrong in placing the stutzeri inside Pseudomonas (as opposed to Pseudomonadaceae) or the taxonomy is wrong.

(Update document 'pg_2448' via OpenTree API)