Rerooted both trees per publication
The article says nothing about ingroup, outgroup, or rooting, but I assume Fig. 2b is rooted correctly, and I assume the root is the ingroup. Neither tree was rooted by the original curator in agreement with the article. I marked the tree corresponding to Fig. 2b as preferred since it is the only one shown in the article. No idea how to resolve duplicates.

(Update document 'pg_108' via OpenTree API)