summary tree added
Currently OTT includes Polypodium in Trachylinae. This tree summarizes the results of this study in a way that should override OT's out-of-date taxonomy. When OTT is fixed we should be able to get rid of this tree. Note that this study's sampling of non-Polypodium Trachylines may not be as good as this tree implies. This is just trying to convey the notion that the other taxa in Trachylina are expected to be monophyletic and sister to Hydroidolinae

(Update document 'ot_519' via OpenTree API)