Uploaded at FuturePhy Chicago, May 5, 2017
Abstract
Phylogenetic relationships of Rhizoctonia fungi within the order Cantharellales were studied using sequence data from portions of the ribosomal DNA cluster regions ITS-LSU, rpb2, tef1, and atp6 for 50 taxa, and public sequence data from the rpb2 locus for 165 taxa. Data sets were analysed individually and combined using Maximum Parsimony, Maximum Likelihood, and Bayesian Phylogenetic Inference methods. All analyses supported the monophyly of the family Ceratobasidiaceae, which comprises the genera Ceratobasidium and Thanatephorus. Multi-locus analysis revealed 10 well-supported monophyletic groups that were consistent with previous separation into anastomosis groups based on hyphal fusion criteria. This analysis coupled with analyses of a larger sample of 165 rpb2 sequences of fungi in the Cantharellales supported a sister relationship between the Botryobasidiaceae and Ceratobasidiaceae and a sister relationship of the Tulasnellaceae with the rest of the Cantharellales. The inclusion of additional sequence data did not clarify incongruences observed in previous studies of Rhizoctonia fungi in the Cantharellales based on analyses of a single or multiple genes. The diversity of ecological and morphological characters associated with these fungi requires further investigation on character evolution for re-evaluating homologous and homoplasious characters.

(Update document 'ot_1088' via OpenTree API)