Discrepancy between File and Published Phylogeny
We uploaded the data as uploaded in TreeBase (double-checked in Mesquite). However, these data are not consistent with the tree provided in the paper. Two of the outgroup species are incorrectly nested inside Sabia and many of the other internal nodes/species placements were incorrect. It is unclear how these inconsistencies arose.