completed curation. just needs a license.
Most of the answers are on p. 684 and p. 687 of the paper.  Scolia and Apterogyna are "outer outgroups" used to root the tree, so I put the root there.  Amblyoponinae is the ingroup declared in the titled and denoted on the tree figure.  Method is maximum likelihood and the tree is time-scaled using BEAST.

(Update document 'ot_1249' via OpenTree API)