Remapped most tips to species
Tips had been mapped to genus. Many of the tips in the tree represent multiple species concatenated into a genus. However, the phylogenetic distance between these tips and the undesired downstream behavior in synthesis of mapping to genus leads me to remap most of them the species most represented in the genus chimera. If the chimera came from multiple genera I did not map it. Some of the taxonomy has changed, e.g. some Pipra are now Lepidothrix, but in all cases there is nothing else in the tree close enough to matter.

(Update document 'ot_757' via OpenTree API)