Initial commit Faille et al. (2010)
FailleRibera2010Trechini.nwk
Confusing taxonomy partly due to use of ambiguated (abbreviated) names, e.g. there is both a G. vandeli and A. vandeli used in the figures, text and sample table. Figure 4 seems to be a subset of Figure 2, but not specified in the legends (both stated as Trechini).

(Update document 'ot_2679' via OpenTree API)